+++ packages/terra-core-docs/src/terra-dev-site/test/tag/OnClickTag.test.jsx
@@ -1,15 +1,19 @@
+/* eslint-disable no-alert */
 import React from 'react';
 import classNames from 'classnames/bind';
-import Tag from '../../../Tag';
+import Tag from 'terra-tag';
 import styles from './TagTestCommon.module.scss';
 
 const cx = classNames.bind(styles);
-const DefaultTag = () => (
+
+const OnClickTag = () => (
   <div className={cx('tag-wrapper')}>
     <Tag
-      id="default"
-      text="Default Tag"
+      id="on-click"
+      text="Clickable Tag"
+      onClick={() => window.alert('Tag has been clicked!')}
     />
   </div>
 );
-export default DefaultTag;
+
+export default OnClickTag;