+++ packages/terra-core-docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Added
+  * Migrate docs from terra-tag.
   * Migrate docs from terra-toggle-button.
   * Migrate docs from terra-visually-hidden-text.
   * Migrate docs from terra-toolbar.